APIs/azure.com/recoveryservicesbackup-bms/2019-06-15/python-aiohttp-client/tests/test_protected_items_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.protected_item_resource import ProtectedItemResource


pytestmark = pytest.mark.asyncio

async def test_protected_items_create_or_update(client):
    """Test case for protected_items_create_or_update

    
    """
    parameters = {"properties":{"deferredDeleteTimeInUTC":"2000-01-23T04:56:07.000+00:00","createMode":"Invalid","backupSetName":"backupSetName","sourceResourceId":"sourceResourceId","backupManagementType":"Invalid","isRehydrate":True,"lastRecoveryPoint":"2000-01-23T04:56:07.000+00:00","isScheduledForDeferredDelete":True,"policyId":"policyId","deferredDeleteTimeRemaining":"deferredDeleteTimeRemaining","containerName":"containerName","protectedItemType":"protectedItemType","isDeferredDeleteScheduleUpcoming":True,"workloadType":"Invalid"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/Subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.RecoveryServices/vaults/{vault_name}/backupFabrics/{fabric_name}/protectionContainers/{container_name}/protectedItems/{protected_item_name}'.format(vault_name='vault_name_example', resource_group_name='resource_group_name_example', subscription_id='subscription_id_example', fabric_name='fabric_name_example', container_name='container_name_example', protected_item_name='protected_item_name_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_protected_items_delete(client):
    """Test case for protected_items_delete

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/Subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.RecoveryServices/vaults/{vault_name}/backupFabrics/{fabric_name}/protectionContainers/{container_name}/protectedItems/{protected_item_name}'.format(vault_name='vault_name_example', resource_group_name='resource_group_name_example', subscription_id='subscription_id_example', fabric_name='fabric_name_example', container_name='container_name_example', protected_item_name='protected_item_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_protected_items_get(client):
    """Test case for protected_items_get

    
    """
    params = [('api-version', 'api_version_example'),
                    ('$filter', 'filter_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/Subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.RecoveryServices/vaults/{vault_name}/backupFabrics/{fabric_name}/protectionContainers/{container_name}/protectedItems/{protected_item_name}'.format(vault_name='vault_name_example', resource_group_name='resource_group_name_example', subscription_id='subscription_id_example', fabric_name='fabric_name_example', container_name='container_name_example', protected_item_name='protected_item_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

